svr: factor auth fb-conf
